xml indenter work your magic

+++ apps/internetfloorplans/templates/xsl/equipment_type_summary.xsl
@@ -21,30 +21,29 @@ along with this program; if not, see http://www.gnu.org/licenses
 or write to the Free Software Foundation,Inc., 51 Franklin Street,
 Fifth Floor, Boston, MA 02110-1301 USA
 -->
-<xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform" >
-<xsl:template name="equipment-type-summary">
-<xsl:for-each select="//get_equipment_type_by_id">
-	<tr>
-		<td>
-		<xsl:value-of select="/_R_/i18n/options"/>:<br />
-		<a href="{//link_prefix}ifp-equipment-type-edit&amp;equipment_type_id={equipment_type_id}">
+<xsl:stylesheet version="1.0" xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
+  <xsl:template name="equipment-type-summary">
+    <xsl:for-each select="//get_equipment_type_by_id">
+      <tr>
+        <td>
+          <xsl:value-of select="/_R_/i18n/options"/>:<br />
+          <a href="{//link_prefix}ifp-equipment-type-edit&amp;equipment_type_id={equipment_type_id}">
             <xsl:value-of select="/_R_/i18n/edit_this_equipment_type"/>
-        </a>
-        <br/>
-		<a href="{//link_prefix}ifp-equipment-type-delete&amp;equipment_type_id={equipment_type_id}"
-            onclick="return confirm('{/_R_/i18n/confirm_delete}')">
+          </a>
+          <br/>
+          <a href="{//link_prefix}ifp-equipment-type-delete&amp;equipment_type_id={equipment_type_id}" onclick="return confirm('{/_R_/i18n/confirm_delete}')">
             <xsl:value-of select="/_R_/i18n/delete_this_equipment_type"/>
-        </a>
-		</td>
-		<td>
-		<xsl:value-of select="/_R_/i18n/name"/>:<br/>
-		<xsl:value-of select="//equipment_type_get_by_id/name" />
-		</td>
-		<td colspan="2">
-		<xsl:value-of select="/_R_/i18n/description"/>: <br/>
-		<xsl:value-of select="//equipment_type_get_by_id/description" />
-		</td>
-	</tr>
-</xsl:for-each>
-</xsl:template>
+          </a>
+        </td>
+        <td>
+          <xsl:value-of select="/_R_/i18n/name"/>:<br/>
+          <xsl:value-of select="//equipment_type_get_by_id/name" />
+        </td>
+        <td colspan="2">
+          <xsl:value-of select="/_R_/i18n/description"/>: <br/>
+          <xsl:value-of select="//equipment_type_get_by_id/description" />
+        </td>
+      </tr>
+    </xsl:for-each>
+  </xsl:template>
 </xsl:stylesheet>
